baseline/pyproject.toml
[project]
name = "baseline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=4.8.1",
    "dotenv>=0.9.9",
    "elevenlabs>=2.13.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.30.1",
    "loguru>=0.7.3",
    "openai>=1.102.0",
    "soundfile>=0.13.1",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]


[tool.black]
line-length = 100

[tool.isort]
profile = "black"
